Move ValidationPipelineBehavior and UnitOfWorkPipelineBehavior to respective namespaces
